Undo previous input component changes and start working on edit cancel
The previous changes seemed like a good way to optimize not
changing state so often, but after experimenting, it seems like
it actually makes sense to do so so that the state changes
can be propagated all at once. The changes made it so that
there was some visible lag when updating.